Refactor client code and test application.
* Move the test application into redis_client module.
* Tweak cli module to prepare for adding sentinel support.
