+++ krux_redis/cli.py
@@ -8,7 +8,9 @@ from __future__ import absolute_import
 from krux.cli import get_group
 
 
-DEFAULT_TIMEOUT = 0.5   # in seconds
+DEFAULT_REDIS_MASTER = 'redis://localhost:6379/0'
+DEFAULT_REDIS_TIMEOUT = 0.5   # in seconds
+GROUP_NAME = 'redis'
 
 
 # Designed to be called from krux.cli, or programs inheriting from it
@@ -16,11 +18,11 @@ DEFAULT_TIMEOUT = 0.5   # in seconds
 # TODO: Potentially refactor this. See: https://git.io/vSFYS
 def add_redis_cli_arguments(parser):
 
-    group = get_group(parser, 'redis')
+    group = get_group(parser, GROUP_NAME)
 
     group.add_argument(
         '--redis-master',
-        default='redis://localhost:6379/0',
+        default=DEFAULT_REDIS_MASTER,
         help='Redis master url (default: %(default)s)',
     )
 
@@ -33,6 +35,6 @@ def add_redis_cli_arguments(parser):
 
     group.add_argument(
         '--redis-timeout',
-        default=DEFAULT_TIMEOUT,
+        default=DEFAULT_REDIS_TIMEOUT,
         help='Timeout for redis calls, in seconds (default: %(default)s)'
     )
